# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/division/addUser": {
            "post": {
                "description": "Add User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "division"
                ],
                "summary": "Add User",
                "parameters": [
                    {
                        "description": "User request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddUserAndRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseOK"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/division/all": {
            "get": {
                "description": "Get Divisions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "division"
                ],
                "summary": "Get Divisions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetDivisions"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/division/create": {
            "post": {
                "description": "Create Division",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "division"
                ],
                "summary": "Create Division",
                "parameters": [
                    {
                        "description": "User request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateDivisionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseOKWithGUID"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/quest/add": {
            "post": {
                "description": "Add Quest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Add Quest",
                "parameters": [
                    {
                        "description": "User request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateQuestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseOK"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/quest/all": {
            "get": {
                "description": "Get All Quest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Get All Quest",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetAllQuestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/quest/stage/add": {
            "post": {
                "description": "Add Stage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quest"
                ],
                "summary": "Add Stage",
                "parameters": [
                    {
                        "description": "User request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateStageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseOK"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "base.Blame": {
            "type": "string",
            "enum": [
                "User",
                "Postgres",
                "Server",
                "Unknown",
                "Neo4j"
            ],
            "x-enum-varnames": [
                "BlameUser",
                "BlamePostgres",
                "BlameServer",
                "BlameUnknown",
                "BlameNeo4j"
            ]
        },
        "base.ResponseFailure": {
            "type": "object",
            "properties": {
                "blame": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/base.Blame"
                        }
                    ],
                    "example": "Guilty System"
                },
                "message": {
                    "type": "string",
                    "example": "error occurred"
                },
                "status": {
                    "type": "string",
                    "example": "Error"
                },
                "trackingID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                }
            }
        },
        "base.ResponseOK": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "trackingID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                }
            }
        },
        "base.ResponseOKWithGUID": {
            "type": "object",
            "properties": {
                "GUID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "trackingID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                }
            }
        },
...
        "model.AddUserAndRoleRequest": {
            "type": "object",
            "properties": {
                "divisionID": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
        "model.CreateDivisionRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.CreateQuestRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "divisionID": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.CreateStageRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "questID": {
                    "type": "string"
                }
            }
        },
...
        "model.DivisionObject": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.GetAllQuestResponse": {
            "type": "object",
            "properties": {
                "quests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.QuestObject"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "trackingID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                }
            }
        },
        "model.GetDivisions": {
            "type": "object",
            "properties": {
                "divisions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DivisionObject"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "trackingID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                }
            }
        },
...
        "model.QuestObject": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StageObject"
                    }
                }
            }
        },
        "model.StageObject": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}